move salesforce files into app
