+++ accounts/tests.py
@@ -8,8 +8,6 @@ from django.conf import settings
 from django.contrib.auth.models import User
 from django.contrib.auth import models
 from django.core.exceptions import ObjectDoesNotExist
-from simple_salesforce import Salesforce as SimpleSalesforce
-from .salesforce import Salesforce
 import unittest
 import shutil
 import os
@@ -75,79 +73,4 @@ import logging
 #         super(LiveServerTestCase, self).tearDown()
 
 
-class SalesforceTest(LiveServerTestCase,WagtailPageTests):
-    def setUp(self):
-        super(WagtailPageTests, self).setUp()
-        super(LiveServerTestCase, self).setUp()
-
-    def test_login(self):
-        sf = SimpleSalesforce(**settings.SALESFORCE)
-        self.assertEqual(sf.sf_instance, u'na12.salesforce.com')
-
-    def test_database_query(self):
-        sf = SimpleSalesforce(**settings.SALESFORCE)
-        contact_info = sf.query(
-            "SELECT Id FROM Contact WHERE Accounts_ID__c = '0'")
-        self.assertEqual(
-            contact_info['records'][0]['Id'], u'003U000001erXyqIAE')
-
-    def test_faculty_confirmed(self):
-        with Salesforce() as sf:
-            status = sf.faculty_status(0)
-            self.assertEqual(status, u'Confirmed')
-
-    def test_faculty_unknown(self):
-        with Salesforce() as sf:
-            status = sf.faculty_status(1)
-            self.assertIsNone(status)
-
-    def test_faculty_pending(self):
-        with Salesforce() as sf:
-            status = sf.faculty_status(2)
-            self.assertEqual(status, u'Pending')
-
-    def test_faculty_rejected(self):
-        with Salesforce() as sf:
-            status = sf.faculty_status(3)
-            self.assertEqual(status, u'Rejected')
-
-    def test_context_manager(self):
-        with open(settings.LOGGING['handlers']['file']['filename'], 'r') as f:
-            lines = f.readlines()
-            if lines:
-                last_message = lines[-1]
-            else:
-                last_message = None
-        with self.assertRaises(RuntimeError):
-            with Salesforce() as sf:
-                raise RuntimeError("test context manager error handling")
-        with open(settings.LOGGING['handlers']['file']['filename'], 'r') as f:
-            lines = f.readlines()
-            new_message = lines[-1]
-        self.assertNotEqual(last_message, new_message)
-        self.assertIn("test context manager error handling", new_message)
-
-    @override_settings(SALESFORCE={})
-    def test_context_manager_handle_init_errors(self):
-        with open(settings.LOGGING['handlers']['file']['filename'], 'r') as f:
-            lines = f.readlines()
-            if lines:
-                last_message = lines[-1]
-            else:
-                last_message = None
-
-        with Salesforce():
-            pass
-
-        with open(settings.LOGGING['handlers']['file']['filename'], 'r') as f:
-            lines = f.readlines()
-            new_message = lines[-1]
-        self.assertNotEqual(last_message, new_message)
-        self.assertIn(
-            'You must provide login information or an instance and token', new_message)
-        responce=self.client.get('/api/user/',follow=True, secure=True)
-        self.assertEqual(responce.status_code,200)
-    def tearDown(self):
-        super(WagtailPageTests, self).tearDown()
-        super(LiveServerTestCase, self).tearDown()
