+++ api/views.py
@@ -2,7 +2,7 @@ from rest_framework import viewsets
 from django.contrib.auth.models import User
 from .serializers import ImageSerializer, UserSerializer
 from wagtail.wagtailimages.models import Image
-from accounts.salesforce import Salesforce
+from salesforce.salesforce import Salesforce
 from django.contrib.auth.models import Group
 from django.conf import settings
